Add smooth damage feedback 💭✨

+++ Assets/3.Scripts/ScriptableObject/Animal/Scripts/AnimalData.cs
@@ -7,8 +7,8 @@ public class AnimalData : ScriptableObject
     public string AnimalName {  get { return animalName; } }
     [SerializeField] private float maxHp;
     public float MaxHP { get { return maxHp; } }
-    [SerializeField] private float damage;
-    public float Damage { get { return damage; } }
+    [SerializeField] private int damage;
+    public int Damage { get { return damage; } }
     [SerializeField] private float attakDistance;
     public float AttackDistance { get { return attakDistance; } }
     [SerializeField] private float findRange;
